refactor: Extract domains from web_fetch, parsing from web_search
- Move PRE_APPROVED_DOMAINS (~100 lines) to web_fetch_domains.rs
- Move web_fetch tests (~240 lines) to web_fetch_tests.rs
- Extract stream event parsing logic into web_search_parse.rs (~160 lines)
  with a standalone parse_search_results() function
- Move web_search tests (~400 lines) to web_search_tests.rs

Results:
  web_fetch.rs: 874 -> 532 LOC
  web_search.rs: 908 -> 395 LOC
  Both now well under the 600 LOC target.
